refactor(tests): unit-tests in separate file
This seems cleaner to be, especially under the light of more complex
tests which are to follow.
